requirement.txt
blinker==1.4
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-sqlacodegen==1.1.6.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.17.1
inflect==2.1.0
itsdangerous==1.1.0
javascripthon==0.10
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
macropy3==1.1.0b2
MarkupSafe==1.1.1
Pillow==5.4.1
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
PyMySQL==0.9.3
SQLAlchemy==1.3.0
Werkzeug==0.14.1
WTForms==2.2.1
